% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_prompt.R
\name{pkg_prompt}
\alias{pkg_prompt}
\title{Convert package documentation to markdown}
\usage{
pkg_prompt(pkg, topics = NULL)
}
\arguments{
\item{pkg}{A single string giving the package name.}

\item{topics}{The topics from the package documentation to include, as listed
in \code{help(package = pkg)}. If \code{NULL}, all package topics will be included in
output.}
}
\value{
A single string giving markdown descriptions for each help topic in the
package. The function will error if the package isn't installed.
}
\description{
Convert package documentation to markdown
}
\examples{
pkg_prompt("splines")

}
