% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgprompt-package.R
\docType{package}
\name{pkgprompt-package}
\alias{pkgprompt}
\alias{pkgprompt-package}
\title{pkgprompt: Convert package documentation to LLM-readable plain text}
\description{
Information in a large language model's prompt is analogous to its short-term memory; models will provide more informative answers on problem contexts that they've been prompted on, even if the same information was in the training data. This package converts R package documentation to LLM-readable plain text to facilitate passing it as part of a prompt.
}
\author{
\strong{Maintainer}: Simon Couch \email{simon.couch@posit.co} (\href{https://orcid.org/0000-0001-5676-5107}{ORCID})

}
\keyword{internal}
